Move opengever.ogds.model.query to opengever.base.
